Added VisualMicro files to .gitignore
VisualMicro (www.visualmicro.com) is a plugin for Visual Studio or Atmel
Studio (which is Atmel's redistribution of Visual Studio), that makes it
possible to use VS to edit, compile and download Arduino sketches. It
generates and uses a number of temporary files and housekeeping files,
which can all be re-generated by opening the .ino file. They are
distracting and aren't used by users who use the Arduino IDE so they
were added to .gitignore.