Added GPS clock and light-sensitive screen saver
The program gets the date and time from the GPS module and shows it on
the screen, along with a rotating display that shows the date,
temperature and measured light from the LDR's. When it's dark (as
determined by an EEPROM-stored threshold) the LCD backlight and RoboBrrd
LEDs are switched off after a certain amount of time.