Added clock, night light, menu
If the RoboBrrd feature is enabled and the RoboBrrd has a GPS module
installed, it will now show a clock with time and date by default.
Pushing the Select button will start the Hunt the Wumpus game.
The RoboBrrd eye LEDs and the LCD backlight serve as a night light. If
the ambient light drops under a certain value, the lights are turned off
after a preset time. If the EEPROM feature is enabled, these settings
can be adjusted in the Settings struct.
If the EEPROM feature is enabled, pushing the button in the Clock menu
starts the menu. I'm not too happy with the menu yet, but it's not done
yet either. Currently the menu can be used to start the game and to
change the backlight color for the night light in the EEPROM. When the
system is reset, the new color setting is automatically retrieved from
the EEPROM by the EEPROM_mgr::Begin() function which is called during
setup().
I also added a menu option to set an alarm but this hasn't been
implemented yet so it has no effect yet.